refactor: rename notifier package and update Room and User classes to extend BaseEntity
