+++ src/main/java/com/icebreaker/be/global/common/entity/UpdatableEntity.java
@@ -1,0 +1,19 @@
+package com.icebreaker.be.global.common.entity;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.annotation.JsonFormat.Shape;
+import jakarta.persistence.Column;
+import jakarta.persistence.MappedSuperclass;
+import java.time.LocalDateTime;
+import lombok.Getter;
+import org.springframework.data.annotation.LastModifiedDate;
+
+@Getter
+@MappedSuperclass
+public abstract class UpdatableEntity extends BaseEntity {
+
+    @LastModifiedDate
+    @Column(nullable = false)
+    @JsonFormat(shape = Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
+    private LocalDateTime updatedAt;
+}
